Add automatic resizing for high DPI (Retina screens)

+++ src/tasks/matchImageSnapshot.js
@@ -4,10 +4,10 @@ const path = require('path');
 const { getConfig } = require('../config');
 const { getImageSnapshotFilename, getImageData } = require('../utils/imageSnapshots');
 const saveImageSnapshot = require('./saveImageSnapshot');
-const { getImageObject, compareImages, moveActualImageToSnapshotsDirectory, createDiffObject } = require('../utils/tasks/imageSnapshots');
-const { IMAGE_TYPE_DIFF } = require('../constants');
+const { getImageObject, compareImages, moveActualImageToSnapshotsDirectory, createDiffObject, resizeImage } = require('../utils/tasks/imageSnapshots');
+const { IMAGE_TYPE_DIFF, IMAGE_TYPE_ACTUAL } = require('../constants');
 
-function matchImageSnapshot(data = {}) {
+async function matchImageSnapshot(data = {}) {
   const {
     commandName,
     dataType,
@@ -17,16 +17,27 @@ function matchImageSnapshot(data = {}) {
     subject,
     testFile,
   } = data;
+  const actualFilename = getImageSnapshotFilename(testFile, snapshotTitle, IMAGE_TYPE_ACTUAL);
   const diffFilename = getImageSnapshotFilename(testFile, snapshotTitle, IMAGE_TYPE_DIFF);
-  moveActualImageToSnapshotsDirectory(data);
-
   const config = merge({}, cloneDeep(getConfig()), options);
   const snapshotFile = getImageSnapshotFilename(testFile, snapshotTitle);
+  const resized = options.imageConfig.resizeDevicePixelRatio && image.devicePixelRatio !== 1;
+  if (resized) {
+    await resizeImage(image.path, actualFilename, image.devicePixelRatio);
+  }
+  if (resized === false) {
+    moveActualImageToSnapshotsDirectory(data);
+  } else {
+    image.path = actualFilename;
+  }
+
   const expected = getImageObject(snapshotFile);
   const exists = expected !== false;
   const autoPassed = (config.autopassNewSnapshots && expected === false);
-  const actual = exists ? getImageObject(image.path) : image;
+  const actual = exists || resized ? getImageObject(image.path, true) : image;
   const passed = expected && compareImages(expected, actual, diffFilename, options);
+
+  actual.resized = resized !== false;
 
   let updated = false;
 
@@ -35,7 +46,7 @@ function matchImageSnapshot(data = {}) {
     updated = true;
   }
 
-  if (passed) {
+  if (passed && actual && actual.path) {
     rimraf(actual.path);
   }
